+++ Backend/ForTea.Core/Psi/Formatting/T4FormatterSettingsKey.cs
@@ -3,7 +3,7 @@ using JetBrains.ReSharper.Psi.CSharp.CodeStyle.FormatSettings;
 using JetBrains.ReSharper.Psi.EditorConfig;
 using JetBrains.ReSharper.Psi.Format;
 
-namespace GammaJul.ForTea.Core.Psi.CodeStyle
+namespace GammaJul.ForTea.Core.Psi.Formatting
 {
 	[SettingsKey(typeof(CSharpFormatSettingsKey), "Code formatting in T4 C#")]
 	[EditorConfigKey("t4")]
